+++ openair1/PHY/defs_gNB.h
@@ -66,7 +66,7 @@ typedef struct {
   /// Frame where current HARQ round was sent
   uint32_t frame;
   /// Interleaver outputs
-  uint8_t *f;
+  uint32_t *f;
   /// LDPC lifting size
   uint32_t Z;
   /// REs unavailable for DLSCH (overlapping with PTRS, CSIRS etc.)
